+++ src/globals.d.ts
@@ -1,0 +1,7 @@
+///<reference types="vscode"/>
+
+type ItemOrItems<T> = T | T[]
+
+//#region vscode
+
+//#endregion vscode